10-ai-agents-with-python-and-langchain/requirements.txt
ipykernel
langchain==0.3.15
langchain-community==0.3.15
langchain-openai==0.3.1
langchain-experimental==0.3.4
langchainhub==0.1.21
python-dotenv==1.0.1
pydantic==2.10.5
beautifulsoup4==4.12.3
# wikipedia==1.4.0
# arxiv==2.1.3
stack-data==0.6.3
pandas==2.2.3
tabulate==0.9.0
stackapi
